+++ buffalo/cmd/app_generators.go
@@ -25,12 +25,12 @@ func newAppGenerator(data gentronics.Data) *gentronics.Generator {
 		RemotePath: "https://raw.githubusercontent.com/gobuffalo/buffalo/master/logo.svg",
 	})
 	g.Add(gentronics.NewFile(".gitignore", nGitignore))
-	g.Add(gentronics.NewCommand(goGet("github.com/markbates/refresh/...")))
-	g.Add(gentronics.NewCommand(goInstall("github.com/markbates/refresh")))
-	g.Add(gentronics.NewCommand(goGet("github.com/markbates/grift/...")))
-	g.Add(gentronics.NewCommand(goInstall("github.com/markbates/grift")))
-	g.Add(gentronics.NewCommand(goGet("github.com/motemen/gore")))
-	g.Add(gentronics.NewCommand(goInstall("github.com/motemen/gore")))
+	g.Add(gentronics.NewCommand(generate.GoGet("github.com/markbates/refresh/...")))
+	g.Add(gentronics.NewCommand(generate.GoInstall("github.com/markbates/refresh")))
+	g.Add(gentronics.NewCommand(generate.GoGet("github.com/markbates/grift/...")))
+	g.Add(gentronics.NewCommand(generate.GoInstall("github.com/markbates/grift")))
+	g.Add(gentronics.NewCommand(generate.GoGet("github.com/motemen/gore")))
+	g.Add(gentronics.NewCommand(generate.GoInstall("github.com/motemen/gore")))
 	g.Add(generate.NewWebpackGenerator(data))
 	g.Add(newSodaGenerator())
 	g.Add(gentronics.NewCommand(appGoGet()))
