+++ buffalo/cmd/root.go
@@ -23,7 +23,9 @@ package cmd
 import (
 	"fmt"
 	"os"
+	"os/exec"
 
+	"github.com/gobuffalo/buffalo/buffalo/cmd/generate"
 	"github.com/spf13/cobra"
 )
 
@@ -45,6 +47,14 @@ func Execute() {
 		fmt.Println(err)
 		os.Exit(-1)
 	}
+}
+
+func goInstall(pkg string) *exec.Cmd {
+	return generate.GoInstall(pkg)
+}
+
+func goGet(pkg string) *exec.Cmd {
+	return generate.GoGet(pkg)
 }
 
 // func init() {
